requirements.txt
requests
ratelimit
pandas
python-Levenshtein